Implementing correct payload memberId when access token is used
